requirements.txt
streamlit==1.24.0
openai==0.27.8
pandas==2.0.3
python-dotenv==1.0.0
matplotlib==3.7.2
Pillow==10.0.0 